test(protocoltest): drop CLI-duplicate e2e wrappers, move unique coverage into the default suite
cli/harness and the e2e-tagged go tests were two front-ends over the same
Matrix engine (Run(t) and ExecuteAll both funnel into executeTest), so the
wrapper files re-ran exactly the cells CI already executes via
harness-matrix.yml - except the wrappers never ran anywhere and rotted.

- Delete the pure wrappers: matrix_test.go (Run wrappers), transitive_test.go,
  idempotent_test.go, harness_test.go, client_gosdk_e2e_test.go. Their cheap
  config-shape assertions (DefaultMatrix, chains, idempotent cases, filters)
  move to matrix_config_test.go in the default suite.
- Un-tag the files whose coverage exists nowhere else so they run in default
  CI: failover_test.go (19 retry/failover semantics tests, no harness leg
  covers failover), endpoint_distinction_test.go, idempotent_mechanism_test.go.
- roundtrip_test.go: remove the 9 happy-path cases that are strict subsets of
  matrix cells; keep and un-tag what the matrix cannot cover - the
  OpenAIChat->AnthropicV1 and V1->V1 pairs (not in DefaultPairs), error-status
  passthrough semantics, the prime-path anchor, and the Codex assembly branch.

internal/protocoltest now has no e2e-tagged files: everything either runs in
the default suite or via the harness CI legs, so nothing is left in the
never-executed dark zone that produced this rot.
